An improved titan :D and libsaturn includes

+++ LibSaturn/Source/Test.c
@@ -1,4 +1,6 @@
-int Add(int a, int b)
+#include "Core.h"
+
+u64 Add(u64 a, u64 b)
 {
 	return a + b;
 }
